refactor: Refactor serialization (#14)
